internationalize toast message & refactor
