+++ src/components/HomePage/PrayerTimes/PrayerTimes.tsx
@@ -2,13 +2,16 @@ import { useEffect } from 'react';
 
 import useTranslation from 'next-translate/useTranslation';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
+import { toast } from 'react-toastify';
 import useSWR from 'swr';
 
-import FindLocationIcon from '../../../public/icons/find-location.svg';
-import { ToastContainer } from '../dls/Toast/Toast';
-import { setAccurateLocation } from '../Navbar/SettingsDrawer/PrayerTimesSection';
+import FindLocationIcon from '../../../../public/icons/find-location.svg';
+import { ToastContainer } from '../../dls/Toast/Toast';
+import { setAccurateLocation } from '../../Navbar/SettingsDrawer/PrayerTimesSection';
 
 import styles from './PrayerTimes.module.scss';
+import { formatLocation, formatTime, getNextPrayerTime } from './PrayerTimesHelper';
+import { PrayerTimesData, HijriDateData } from './PrayerTimesTypes';
 
 import { fetcher } from 'src/api';
 import Button, { ButtonType, ButtonVariant } from 'src/components/dls/Button/Button';
@@ -22,7 +25,7 @@ import {
 import { makePrayerTimesUrl } from 'src/utils/apiPaths';
 
 const PrayerTimes = () => {
-  const { t } = useTranslation('home');
+  const { t } = useTranslation('');
   const geoLocation = useSelector(selectGeoLocation, shallowEqual);
   const geoPermission = useSelector(selectGeoPermission);
   const dispatch = useDispatch();
@@ -30,11 +33,11 @@ const PrayerTimes = () => {
   useEffect(() => {
     // get the location every time the user refresh the page if we already have the permission
     if (geoPermission === GeoPermission.Granted) setAccurateLocation(dispatch);
-  }, [dispatch, geoPermission]);
+  }, [dispatch, geoPermission, t]);
 
   const calculationMethod = useSelector(selectCalculationMethod);
   const madhab = useSelector(selectMadhab);
-  const { data } = useSWR<Data>(
+  const { data } = useSWR<PrayerTimesData>(
     makePrayerTimesUrl({
       calculationMethod,
       madhab,
@@ -56,7 +59,11 @@ const PrayerTimes = () => {
         <div className={styles.prayerTimesContainer}>
           <div className={styles.locationContainer}>
             <Button
-              onClick={() => setAccurateLocation(dispatch)}
+              onClick={() =>
+                setAccurateLocation(dispatch, () =>
+                  toast(t('common:prayer-times.location-updated')),
+                )
+              }
               type={ButtonType.Secondary}
               className={styles.findLocationButton}
               variant={ButtonVariant.Ghost}
@@ -68,7 +75,7 @@ const PrayerTimes = () => {
           {nextPrayerTime && (
             <div>
               <span className={styles.prayerName}>
-                {t(`prayer-names.${nextPrayerTime.prayerName}`)}
+                {t(`home:prayer-names.${nextPrayerTime.prayerName}`)}
               </span>{' '}
               <span>
                 {formatTime(nextPrayerTime.time.getHours())}:
@@ -81,37 +88,6 @@ const PrayerTimes = () => {
       <ToastContainer />
     </>
   );
-};
-
-type PrayerTimes = {
-  fajr: string;
-  sunrise: string;
-  dhuhr: string;
-  asr: string;
-  maghrib: string;
-  isha: string;
-};
-
-type Geo = {
-  city?: string;
-  country?: string;
-  region?: string;
-  latitude?: string;
-  longitude?: string;
-};
-
-type HijriDateData = {
-  dayName: string;
-  month: number;
-  date: number;
-  year: number;
-};
-
-type Data = {
-  geo: Geo;
-  prayerTimes: PrayerTimes;
-  hijriDate: string;
-  hijriDateData: HijriDateData;
 };
 
 const useHijriDateFormatter = (hijriDate?: HijriDateData) => {
@@ -127,49 +103,6 @@ const useHijriDateFormatter = (hijriDate?: HijriDateData) => {
     month,
     year: hijriDate.year,
   });
-};
-
-const formatTime = (time: number) => time.toString().padStart(2, '0');
-const formatLocation = (geo: Geo) => {
-  const location = [];
-  if (geo.city) {
-    location.push(geo.city);
-  }
-  if (geo.country) {
-    location.push(geo.country);
-  }
-
-  return location.join(', ');
-};
-
-const getNextPrayerTime = (
-  prayerTimes: PrayerTimes,
-): {
-  prayerName: string;
-  time: Date;
-} => {
-  const now = new Date();
-
-  const prayerTimeEntries = Object.entries(prayerTimes).sort((a, b) => {
-    const timeA = new Date(a[1]);
-    const timeB = new Date(b[1]);
-    return timeA.getTime() - timeB.getTime();
-  });
-
-  let nextPrayerTime = prayerTimeEntries.find((prayerTime) => {
-    const [, time] = prayerTime;
-    const date = new Date(time);
-    return now < date;
-  });
-
-  // if nextPrayerTime is not found for this day, this means isha is done. So we use fajr as nextPrayerTime
-  if (!nextPrayerTime) nextPrayerTime = prayerTimeEntries[0];
-
-  const [prayerName, time] = nextPrayerTime;
-  return {
-    prayerName,
-    time: new Date(time),
-  };
 };
 
 export default PrayerTimes;
