+++ src/components/HomePage/PrayerTimes/PrayerTimesHelper.ts
@@ -1,0 +1,44 @@
+import { Geo, PrayerTimes } from './PrayerTimesTypes';
+
+export const formatTime = (time: number) => time.toString().padStart(2, '0');
+export const formatLocation = (geo: Geo) => {
+  const location = [];
+  if (geo.city) {
+    location.push(geo.city);
+  }
+  if (geo.country) {
+    location.push(geo.country);
+  }
+
+  return location.join(', ');
+};
+
+export const getNextPrayerTime = (
+  prayerTimes: PrayerTimes,
+): {
+  prayerName: string;
+  time: Date;
+} => {
+  const now = new Date();
+
+  const prayerTimeEntries = Object.entries(prayerTimes).sort((a, b) => {
+    const timeA = new Date(a[1]);
+    const timeB = new Date(b[1]);
+    return timeA.getTime() - timeB.getTime();
+  });
+
+  let nextPrayerTime = prayerTimeEntries.find((prayerTime) => {
+    const [, time] = prayerTime;
+    const date = new Date(time);
+    return now < date;
+  });
+
+  // if nextPrayerTime is not found for this day, this means isha is done. So we use fajr as nextPrayerTime
+  if (!nextPrayerTime) nextPrayerTime = prayerTimeEntries[0];
+
+  const [prayerName, time] = nextPrayerTime;
+  return {
+    prayerName,
+    time: new Date(time),
+  };
+};
